Getting values of rabbit/redis host/port directly from config on sensor instantiation. Removed useless routing_args parameter on BaseRMQSensor. Changes applied also to RMQPublisher in sy.api.
This was accomplished in order not to make sensor instantiation more complicated
from the daemon. Sensor seem not to be such configurable... In any case, in
the future, instead of launching a sensor with custom init_args, we could
create different sensors inheriting from base classes.
Services host and port are not supposed to be dynamic, but can be taken
directly from config without losing flexibility.

Not-so-relevant fixes in config module.